Reference

Contents

Reference

Contents

Index

Index

COPIERTemplate.COPIERTemplate

COPIERTemplate.Copier

Base.copy

COPIERTemplate.Copier.recopy

COPIERTemplate.Copier.update

COPIERTemplate._read_data_from_existing_path

COPIERTemplate._sanitize_package_name

COPIERTemplate.generate

COPIERTemplate.update

COPIERTemplate.COPIERTemplate — Module

COPIERTemplate.jl

This package defines a copier template for Julia packages and a basic user interface aroud copier to use it from Julia.

The main functions are: generate

COPIERTemplate._read_data_from_existing_path — Method

data = _read_data_from_existing_path(dst_path)

Reads the destination folder to figure out some answers.

COPIERTemplate._sanitize_package_name — Method

package_name = _sanitize_package_name(path)

Sanitize the path to guess the package_name by looking at the base name and removing an extension. If the result is not a valid identifier, returns "".

COPIERTemplate.generate — Function

generate(dst_path[, data]; kwargs...)
generate(src_path, dst_path[, data]; true, kwargs...)

Runs the copy command of copier with the COPIERTemplate template. If src_path is not informed, the GitHub URL of COPIERTemplate.jl is used.

The data argument is a dictionary of answers (values) to questions (keys) that can be used to bypass some of the interactive questions.

Keyword arguments

warn_existing_pkg::Boolean = true: Whether to check if you actually meant update. If you run generate and the dst_path contains a .copier-answers.yml, it means that the copy was already made, so you might have means update instead. When true, a warning is shown and execution is stopped.

The other keyword arguments are passed directly to the internal Copier.copy.

COPIERTemplate.update — Function

update([data]; kwargs...)
update(dst_path[, data]; kwargs...)

Run the update command of copier, updating the dst_path (or the current path if omitted) with a new version of the template with a new version of the template.

The data argument is a dictionary of answers (values) to questions (keys) that can be used to bypass some of the interactive questions.

Keyword arguments

The keyword arguments are passed directly to the internal Copier.update.

COPIERTemplate.Copier — Module

Wrapper around copier.

There are three functions in this module:

copy

recopy

update

Base.copy — Function

copy(dst_path[, data]; kwargs...)
copy(src_path, dst_path[, data]; kwargs...)

Wrapper around copier.run_copy.

This is an internal function, if COPIERTemplate's main API is not sufficient, open an issue.

COPIERTemplate.Copier.recopy — Function

recopy(dst_path[, data]; kwargs...)

Wrapper around copier.run_recopy.

This is an internal function, if COPIERTemplate's main API is not sufficient, open an issue.

COPIERTemplate.Copier.update — Function

update(dst_path[, data]; kwargs...)

Wrapper around copier.run_update.

This is an internal function, if COPIERTemplate's main API is not sufficient, open an issue.