Reference

Contents

Reference

Contents

Index

Index

BestieTemplate.BestieTemplate

BestieTemplate.Copier

Base.copy

BestieTemplate.Copier.recopy

BestieTemplate.Copier.update

BestieTemplate._copy

BestieTemplate._read_data_from_existing_path

BestieTemplate._sanitize_package_name

BestieTemplate.apply

BestieTemplate.generate

BestieTemplate.update

BestieTemplate.BestieTemplate — Module

BestieTemplate.jl

This package defines a copier template for Julia packages and a basic user interface aroud copier to use it from Julia.

The main functions are: generate, apply, and update.

BestieTemplate._copy — Method

_copy(src_path, dst_path, data; kwargs...)

Internal function to run common code for new or existing packages.

BestieTemplate._read_data_from_existing_path — Method

data = _read_data_from_existing_path(dst_path)

Reads the destination folder to figure out some answers.

BestieTemplate._sanitize_package_name — Method

package_name = _sanitize_package_name(path)

Sanitize the path to guess the package_name by looking at the base name and removing an extension. If the result is not a valid identifier, returns "".

BestieTemplate.apply — Function

apply(dst_path[, data]; kwargs...)
apply(src_path, dst_path[, data]; true, kwargs...)

Applies the template to an existing project at path $dst_path$. If the dst_path does not exist, this will throw an error. For new packages, use BestieTemplate.generate instead.

Runs the copy command of copier with the BestieTemplate template. If src_path is not informed, the GitHub URL of BestieTemplate.jl is used.

The data argument is a dictionary of answers (values) to questions (keys) that can be used to bypass some of the interactive questions.

Keyword arguments

warn_existing_pkg::Boolean = true: Whether to check if you actually meant update. If you run apply and the dst_path contains a .copier-answers.yml, it means that the copy was already made, so you might have means update instead. When true, a warning is shown and execution is stopped.

The other keyword arguments are passed directly to the internal Copier.copy.

BestieTemplate.generate — Function

generate(dst_path[, data]; kwargs...)
generate(src_path, dst_path[, data]; true, kwargs...)

Generates a new project at the path dst_path using the template. If the dst_path already exists, this will throw an error, unless dst_path = ".". For existing packages, use BestieTemplate.apply instead.

Runs the copy command of copier with the BestieTemplate template. If src_path is not informed, the GitHub URL of BestieTemplate.jl is used.

The data argument is a dictionary of answers (values) to questions (keys) that can be used to bypass some of the interactive questions.

Keyword arguments

warn_existing_pkg::Boolean = true: Whether to check if you actually meant update. If you run generate and the dst_path contains a .copier-answers.yml, it means that the copy was already made, so you might have means update instead. When true, a warning is shown and execution is stopped.

The other keyword arguments are passed directly to the internal Copier.copy.

BestieTemplate.update — Function

update([data]; kwargs...)
update(dst_path[, data]; kwargs...)

Run the update command of copier, updating the dst_path (or the current path if omitted) with a new version of the template with a new version of the template.

The data argument is a dictionary of answers (values) to questions (keys) that can be used to bypass some of the interactive questions.

Keyword arguments

The keyword arguments are passed directly to the internal Copier.update.

BestieTemplate.Copier — Module

Wrapper around copier.

There are three functions in this module:

copy

recopy

update

Base.copy — Function

copy(dst_path[, data]; kwargs...)
copy(src_path, dst_path[, data]; kwargs...)

Wrapper around copier.run_copy.

This is an internal function, if BestieTemplate's main API is not sufficient, open an issue.

BestieTemplate.Copier.recopy — Function

recopy(dst_path[, data]; kwargs...)

Wrapper around copier.run_recopy.

This is an internal function, if BestieTemplate's main API is not sufficient, open an issue.

BestieTemplate.Copier.update — Function

update(dst_path[, data]; kwargs...)

Wrapper around copier.run_update.

This is an internal function, if BestieTemplate's main API is not sufficient, open an issue.