Reference

Contents

Reference

Contents

Index

Index

COPIERTemplate.COPIERTemplate

Base.copy

COPIERTemplate.generate

COPIERTemplate.recopy

COPIERTemplate.update

COPIERTemplate.COPIERTemplate — Module

COPIERTemplate.jl

This package defines a copier template for Julia packages and a basic user interface aroud copier to use it from Julia.

The main functions are: generate

Base.copy — Function

copy(dst_path[, data]; kwargs...)
copy(src_path, dst_path[, data]; kwargs...)

Wrapper around Python's copier.run_copy. Use generate for a more user-friendly function.

COPIERTemplate.generate — Function

generate(dst_path[, data]; kwargs...)
generate(src_path, dst_path[, data]; true, kwargs...)

Runs the copy command of copier with the COPIERTemplate template. If src_path is not informed, the GitHub URL of COPIERTemplate.jl is used.

The data argument is a dictionary of answers (values) to questions (keys) that can be used to bypass some of the interactive questions.

Keyword arguments

The keyword arguments are passed directly to copy.

COPIERTemplate.recopy — Function

recopy(dst_path[, data]; kwargs...)

Wrapper around Python's copier.run_recopy.

COPIERTemplate.update — Function

update(dst_path[, data]; kwargs...)

Wrapper around Python's copier.run_update.